requirements.txt
pytorch_lightning==1.3.2
transformers==4.6.0
Pillow==9.1.1
tqdm==4.64.0
numpy==1.22.4
einops==0.4.1
pyarrow
pandas==1.4.2
